+++ Back/ATV-Backend/office-service/axity.office-persistence/src/main/java/com/axity/office/persistence/ProductDocumentRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductDocumentRepository extends JpaRepository<ProductDocument, Integer> {
+public interface ProductDocumentRepository extends JpaRepository<ProductDocument, Long> {
 }
